backend_cms/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-grappelli==4.0.2
django-restframework==0.0.1
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
django-grappelli>=3.1.5
